+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ## [Unreleased]
 
+## [1.0.7] - 2025-01-29
+- Add built-in macros for converting a string to bytes and push it to the stack.
+  - `__BYTES("hello")` -> `PUSH5 0x68656c6c6f`
+  - This can also be used here: `__RIGHTPAD(__BYTES("hello"))`.
+
 ## [1.0.6] - 2025-01-28
 - Allow to use `--debug` for reverting contracts.
 - Refactored parsing for ABI and build-in argument.